Cleanup old usages of Scope
